renamed channel utils
